Finally tracked the problem down: that ONE line in Integer.cpp (and I changed two others for the same reason). As such, I reverted to my home-rolled reference tracking.
The problem was that Integer explicitly called the destructor for BitField. As such, BitField is actually destroyed twice in the destructor. This would be fine, as BitField's destructor leaves the object in a consistent state, except GCC dead-store-eliminates setting the BitHolder pointer to NULL. The double destruction leads to a double free, and thus the crashing.